+++ R/constants.R
@@ -2,4 +2,3 @@ kOldPipelinesSection = "pipelines_dir"
 kRelativeToOutputDirMetadataSections = c("results_subdir", "submission_subdir")
 
 kRelativeToConfigSections = c("metadata", "pipeline_config")
-