+++ tests/testthat/test_project_operations.R
@@ -15,9 +15,9 @@ p = Project(
     package = "pepr"
   )
 )
-p_file_missing = p
-p_file_missing@config$metadata$sample_table = "missing"
-p_subproj1 = Project(
+pFileMissing = p
+pFileMissing@config$metadata$sample_table = "missing"
+pSubproj1 = Project(
   file = system.file(
     "extdata",
     paste0("example_peps-",branch),
@@ -26,7 +26,7 @@ p_subproj1 = Project(
     package = "pepr"
   )
 )
-p_subproj2 = Project(
+pSubproj2 = Project(
   file = system.file(
     "extdata",
     paste0("example_peps-",branch),
@@ -35,7 +35,7 @@ p_subproj2 = Project(
     package = "pepr"
   )
 )
-p_sub = Project(
+pSub = Project(
   file = system.file(
     "extdata",
     paste0("example_peps-",branch),
@@ -44,7 +44,7 @@ p_sub = Project(
     package = "pepr"
   )
 )
-p_implied = Project(
+pImplied = Project(
   file = system.file(
     "extdata",
     paste0("example_peps-",branch),
@@ -52,6 +52,16 @@ p_implied = Project(
     "project_config.yaml",
     package = "pepr"
   )
+)
+
+pDerived = Project(
+    file = system.file(
+        "extdata",
+        paste0("example_peps-",branch),
+        "example_derived",
+        "project_config.yaml",
+        package = "pepr"
+    )
 )
 
 # tests -------------------------------------------------------------------
@@ -62,7 +72,7 @@ test_that("getSubsample method throws errors", {
 })
 
 test_that("getSubsample method returns a correct size DF", {
-  expect_equal(dim(getSubsample(p_sub, "frog_1", "sub_a")), c(1, 4))
+  expect_equal(dim(getSubsample(pSub, "frog_1", "sub_a")), c(1, 4))
 })
 
 test_that(".loadSampleAnnotation returns a Project object", {
@@ -70,38 +80,38 @@ test_that(".loadSampleAnnotation returns a Project object", {
 })
 
 test_that(".loadSampleAnnotation thorws an error when file not fond", {
-  expect_error(.loadSampleAnnotation(p_file_missing))
+  expect_error(.loadSampleAnnotation(pFileMissing))
 })
 
 test_that(".loadSamplesubnnotation always returns a Project", {
   expect_is(.loadSampleSubannotation(p), 'Project')
-  expect_is(.loadSampleSubannotation(p_subproj2), 'Project')
-  expect_is(.loadSampleSubannotation(p_sub), 'Project')
+  expect_is(.loadSampleSubannotation(pSubproj2), 'Project')
+  expect_is(.loadSampleSubannotation(pSub), 'Project')
 })
 
-test_that(".implyColumns returns Project object", {
-  expect_is(.implyAttributes(p_implied), 'Project')
+test_that(".implyAttributes returns Project object", {
+  expect_is(.implyAttributes(pImplied), 'Project')
 })
 
-test_that(".implyColumns returns Project object", {
-  expect_is(.deriveAttributes(p), 'Project')
+test_that(".deriveAttributes returns Project object", {
+  expect_is(.deriveAttributes(pDerived), 'Project')
 })
 
 test_that(".listSubprojects internal function returns correct object type, length and throws errors",
           {
-            expect_equal(length(.listSubprojects(p_subproj1@config)), 2)
-            expect_is(.listSubprojects(p_subproj2@config), 'character')
+            expect_equal(length(.listSubprojects(pSubproj1@config)), 2)
+            expect_is(.listSubprojects(pSubproj2@config), 'character')
             expect_null(.listSubprojects(p@config))
             expect_error(.listSubprojects(1))
           })
 
 test_that("listSubprojects exported method returns correct object type, length and throws errors",
           {
-            expect_equal(length(listSubprojects(p_subproj1)), 2)
-            expect_is(listSubprojects(p_subproj1), 'character')
+            expect_equal(length(listSubprojects(pSubproj1)), 2)
+            expect_is(listSubprojects(pSubproj1), 'character')
             expect_null(listSubprojects(p))
             expect_error(listSubprojects(1))
-            expect_equal(length(listSubprojects(p_subproj1)), 2)
+            expect_equal(length(listSubprojects(pSubproj1)), 2)
           })
 
 test_that("checkSection returns a correct type", {
@@ -112,4 +122,40 @@ test_that("checkSection returns a correct type", {
 test_that("checkSection returns correct value", {
   expect_equal(checkSection(config(p),c("metadata","sample_table")), T)
   expect_equal(checkSection(config(p),c("test")), F)
+})
+
+test_that("activateSubproject does not fail and throws a warning when called 
+          with invalid subproject name", {
+              expect_warning(activateSubproject(pSubproj1, "test"))
+})
+
+test_that("activateSubproject returns a correct object type", {
+    expect_is(activateSubproject(pSubproj1, "newLib2"), "Project")
+})
+
+test_that(".listSubprojects works with different styles of printing", {
+    expect_message(.listSubprojects(config(pSubproj1)))
+    expect_output(.listSubprojects(config(pSubproj1),style="cat"))
+})
+
+test_that("show methods work", {
+    expect_output(show(p))
+    expect_output(show(config(p)))
+})
+
+test_that("getSample errors", {
+    expect_error(getSample(p, "test"))
+})
+
+test_that("getSample returns a correct object type", {
+    expect_is(getSample(p,"frog_1"), "data.table")
+})
+
+test_that("getSample errors", {
+    expect_error(getSubsample(p, "test","test"))
+    expect_error(getSubsample(pSub,"frog_1", "test"))
+})
+
+test_that("getSample returns a correct object type", {
+    expect_is(getSubsample(pSub,"frog_1", "sub_a"), "data.table")
 })